requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
sqlalchemy[asyncio]==2.0.37
aiosqlite==0.20.0
alembic==1.14.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==9.0.0
redis==5.2.1
loguru==0.7.3
numpy==2.2.2
opencv-python-headless==4.11.0.86
sentence-transformers==3.3.1
scikit-learn==1.6.1
openai==1.60.1
pytest==8.3.4
pytest-asyncio==0.25.2
respx==0.22.0
ruff==0.9.2
